Adding basic layout
I have added a basic layout for each component of my portfolio.
As well as adding a footer with social icons which will need to be attached to my pages once created (purely for design at the moment).
I will need to change the layout style on my about page as images seem too big when viewed on a desktop screen.